+++ data/constants.py
@@ -1,6 +1,9 @@
 import os
 from pathlib import Path
 from datetime import datetime
+from dotenv import load_dotenv
+
+load_dotenv()
 
 PROJECT_FOLDER = Path(__file__).resolve().parent.parent
 
@@ -22,4 +25,10 @@ ADB_DEVICE_IP = "192.168.1.160"
 
 # Caching
 CACHING_FOLDER = os.path.join(os.path.expanduser("~"), ".cache/stewart")
+
+# Weather
+OPENWEATHER_API_KEY = os.getenv("OPENWEATHER_API_KEY")
+OPENWEATHER_API = "https://api.openweathermap.org/data/2.5/weather"
+MY_CITY_LAT = os.getenv("MY_CITY_LAT")
+MY_CITY_LON = os.getenv("MY_CITY_LON")